refactor: Refactoring Large File Optimization
Split aiagentservice.ts and aiagentui.ts into smaller, focused modules to improve maintainability and enforce single responsibility principle.

Split files:

1. aiagentservice.ts ->
    - moderate-content.ts: Content moderation logic
    - prompt.ts: Prompt generation and handling
    - check-model.ts: Model validation
    - process-content.ts: Content processing utilities
2. aiagentui.ts ->
    - ai-agent-button.ts: AI command button UI
    - ai-agent-tone-button.ts: Tone selection UI
    - ai-agent-ui-schema.ts: Schema registration

Closes #129
